requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langchain-community>=0.3.0
pinecone-client>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
